# save-backend/backend-api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SAVE Backend API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5800",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/contests/{contestName}/eligible-projects": {
      "get": {
        "description": "Get list of user\u0027s projects that can participate in a given contest.",
        "operationId": "getAvailableProjectsForContest",
        "parameters": [
          {
            "description": "name of a contest",
            "in": "path",
            "name": "contestName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "example": "basic",
            "in": "header",
            "name": "X-Authorization-Source",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "description": "Successfully fetched projects available for contest."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Get projects that can participate in contest.",
        "tags": [
          "contests"
        ]
      }
    },
    "/api/v1/contests/{contestName}/enroll": {
      "get": {
        "description": "Register your public project for a contest.",
        "operationId": "enrollForContest",
        "parameters": [
          {
            "description": "name of a contest",
            "in": "path",
            "name": "contestName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "name of a project",
            "in": "query",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "name of an organization",
            "in": "query",
            "name": "organizationName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "example": "basic",
            "in": "header",
            "name": "X-Authorization-Source",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successfully enrolled for a contest."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Not enough permissions to enroll for a contest with given project."
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Either given project or given contest was not found."
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Only public projects can participate in contests."
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Register for a contest.",
        "tags": [
          "contests"
        ]
      }
    },
    "/api/v1/contests/{contestName}/executions/{organizationName}/{projectName}": {
      "get": {
        "description": "Get list of execution of a project with given name in contest with given name.",
        "operationId": "getContestExecutionsForProject",
        "parameters": [
          {
            "description": "name of an organization",
            "in": "path",
            "name": "contestName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "name of an organization",
            "in": "path",
            "name": "organizationName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "name of a project",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExecutionDto"
                  }
                }
              }
            },
            "description": "Successfully fetched latest project execution in contest."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExecutionDto"
                  }
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExecutionDto"
                  }
                }
              }
            },
            "description": "Either contest is not found or project is not found or execution is not found."
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Get project executions in contest.",
        "tags": [
          "contests"
        ]
      }
    },
    "/api/v1/contests/{contestName}/executions/{organizationName}/{projectName}/latest": {
      "get": {
        "description": "Get latest execution of a project with given name in contest with given name.",
        "operationId": "getLatestExecutionOfProjectInContest",
        "parameters": [
          {
            "description": "name of an organization",
            "in": "path",
            "name": "organizationName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "name of a project",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "name of an organization",
            "in": "path",
            "name": "contestName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "example": "basic",
            "in": "header",
            "name": "X-Authorization-Source",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionDto"
                }
              }
            },
            "description": "Successfully fetched latest project execution in contest."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionDto"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionDto"
                }
              }
            },
            "description": "Either contest is not found or project is not found or execution is not found."
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Get latest project execution in contest.",
        "tags": [
          "contests"
        ]
      }
    },
...
    "/api/v1/contests/{contestName}/my-results": {
      "get": {
        "description": "Get list of best results of your projects in a given contest.",
        "operationId": "getBestResultsInUserProjects",
        "parameters": [
          {
            "description": "name of a contest",
            "in": "path",
            "name": "contestName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "example": "basic",
            "in": "header",
            "name": "X-Authorization-Source",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContestResult"
                  }
                }
              }
            },
            "description": "Successfully fetched your best results."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContestResult"
                  }
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContestResult"
                  }
                }
              }
            },
            "description": "Either given project or given contest was not found."
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Get your best results in contest.",
        "tags": [
          "contests"
        ]
      }
    },
...
    "/api/v1/contests/{contestName}/scores": {
      "get": {
        "description": "Get scores of all projects in contest.",
        "operationId": "getRatingsInContest",
        "parameters": [
          {
            "description": "name of a contest",
            "in": "path",
            "name": "contestName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContestResult"
                  }
                }
              }
            },
            "description": "Successfully fetched contest scores."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContestResult"
                  }
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Get scores of all projects in contest.",
        "tags": [
          "contests"
        ]
      }
    },
...
    "/api/v1/contests/{organizationName}/{projectName}/best": {
      "get": {
        "description": "Get list of contests in which given project has higher results.",
        "operationId": "getBestProjectContests",
        "parameters": [
          {
            "description": "name of an organization",
            "in": "path",
            "name": "organizationName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "name of a project",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "number of contests that will be fetched, default is 4",
            "in": "path",
            "name": "amount",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 4
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContestResult"
                  }
                }
              }
            },
            "description": "Successfully fetched best project contests."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContestResult"
                  }
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Get best contests of a project.",
        "tags": [
          "contests"
        ]
      }
    },
    "/api/v1/contests/{organizationName}/{projectName}/eligible-contests": {
      "get": {
        "description": "Get list of contest names that a given project can participate in.",
        "operationId": "getAvailableContestsForProject",
        "parameters": [
          {
            "description": "name of an organization",
            "in": "path",
            "name": "organizationName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "name of a project",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "example": "basic",
            "in": "header",
            "name": "X-Authorization-Source",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "description": "Successfully fetched contests available for project."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Get contests that can be participated.",
        "tags": [
          "contests"
        ]
      }
    },
...
    "/api/v1/users/{userName}": {
      "get": {
        "operationId": "findByName",
        "parameters": [
          {
            "in": "path",
            "name": "userName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfo"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "users-details-controller"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContestResult": {
        "required": [
          "contestName",
          "hasFailedTest",
          "organizationName",
          "projectName",
          "sdk",
          "submissionStatus"
        ],
        "type": "object",
        "properties": {
          "contestName": {
            "type": "string"
          },
          "hasFailedTest": {
            "type": "boolean"
          },
          "organizationName": {
            "type": "string"
          },
          "projectName": {
            "type": "string"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "sdk": {
            "type": "string"
          },
          "submissionStatus": {
            "type": "string",
            "enum": [
              "ERROR",
              "FINISHED",
              "OBSOLETE",
              "PENDING",
              "RUNNING"
            ]
          },
          "submissionTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ExecutionDto": {
        "required": [
          "allTests",
          "expectedChecks",
          "failedTests",
          "id",
          "matchedChecks",
          "passedTests",
          "runningTests",
          "skippedTests",
          "startTime",
          "status",
          "type",
          "unexpectedChecks",
          "unmatchedChecks"
        ],
        "type": "object",
        "properties": {
          "allTests": {
            "type": "integer",
            "format": "int64"
          },
          "endTime": {
            "type": "integer",
            "format": "int64"
          },
          "expectedChecks": {
            "type": "integer",
            "format": "int64"
          },
          "failedTests": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "matchedChecks": {
            "type": "integer",
            "format": "int64"
          },
          "passedTests": {
            "type": "integer",
            "format": "int64"
          },
          "runningTests": {
            "type": "integer",
            "format": "int64"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "skippedTests": {
            "type": "integer",
            "format": "int64"
          },
          "startTime": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "enum": [
              "ERROR",
              "FINISHED",
              "OBSOLETE",
              "PENDING",
              "RUNNING"
            ]
          },
          "testSuiteSourceName": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "CONTEST_MODE",
              "PRIVATE_TESTS",
              "PUBLIC_TESTS"
            ]
          },
          "unexpectedChecks": {
            "type": "integer",
            "format": "int64"
          },
          "unmatchedChecks": {
            "type": "integer",
            "format": "int64"
          },
          "version": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "basic": {
        "scheme": "basic",
        "type": "http"
      }
    }
  }
}